feat: continue complete js api

+++ polodb.js/polodb.h
@@ -55,7 +55,7 @@ struct DbValue* PLDB_mk_double(double value);
 
 struct DbValue* PLDB_mk_bool(int bl);
 
-struct DbValue* PLDB_mk_int(int value);
+struct DbValue* PLDB_mk_int(long long value);
 
 struct DbValue* PLDB_mk_str(char* content);
